+++ src/index.ts
@@ -10,16 +10,16 @@ import * as Utilities from './Utilities/index'
 import * as SceneTree from './SceneTree/index'
 import * as Renderer from './Renderer/index'
 
+const libsRegistry = new LibsRegistry(packageJson.version)
 const ZeaEngine = {
-  SystemDesc: SystemDesc,
+  libsRegistry,
+  SystemDesc,
   Registry,
   ...Math,
   ...Utilities,
   ...SceneTree,
   ...Renderer
 }
-
-const libsRegistry = new LibsRegistry(packageJson.version)
 
 zeaDebug('Zea Engine version %s', packageJson.version)